% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aws.athena-package.R
\docType{package}
\name{aws.athena-package}
\alias{aws.athena}
\alias{aws.athena-package}
\title{aws.athena: 'AWS Athena' Client Package}
\description{
Client package for 'AWS Athena' <https://aws.amazon.com/athena/>, an interactive query service that makes it easy to analyze data in 'Amazon S3' using standard 'SQL'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ndiquattro/aws.athena}
  \item Report bugs at \url{https://github.com/ndiquattro/aws.athena/issues}
}

}
\author{
\strong{Maintainer}: Nick DiQuattro \email{ndiquattro@gmail.com}

Authors:
\itemize{
  \item Thomas J. Leeper \email{thosjleeper@gmail.com} (\href{https://orcid.org/0000-0003-4097-6326}{ORCID})
}

}
\keyword{internal}
