% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_executions.R
\name{list_athena_executions}
\alias{list_athena_executions}
\alias{get_athena_execution}
\alias{start_athena_execution}
\alias{stop_athena_execution}
\title{AWS Athena Executions}
\usage{
list_athena_executions(n = 50, token = NULL, workgroup = NULL, ...)

get_athena_execution(id, ...)

start_athena_execution(
  query,
  workgroup = NULL,
  database = NULL,
  output = NULL,
  encryption = NULL,
  ...
)

stop_athena_execution(id, ...)
}
\arguments{
\item{n}{An integer specifying the maximum number of results to return.}

\item{token}{A character string specifying a continuation token, for pagination.}

\item{workgroup}{A character string containing the workgroup's name.}

\item{\dots}{Additional arguments passed to \code{\link{athenaHTTP}}}

\item{id}{A character string containing an Amazon Athena Query Executionn ID. A vector of IDs can also be specified to request multiple query executions.}

\item{database}{Name of the database to associate with the query.}
}
\description{
Get, Start, and Stop Athena Query Executions
}
\examples{
\dontrun{
require("aws.s3")
b <- aws.s3::put_bucket("aws-athena-r-demo")

# create an Athena database
id <- start_athena_execution(
  query = "create database DEMO",
  output = "s3://aws-athena-r-demo"
)

# create a table in the database
}
}
\references{
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_ListQueryExecutions.html}{API Reference: ListQueryExecutions}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_GetQueryExecution.html}{API Reference: GetQueryExecution}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_BatchGetQueryExecution.html}{API Reference: BatchGetQueryExecution}
}
