% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_queries.R
\name{list_athena_queries}
\alias{list_athena_queries}
\alias{get_athena_query}
\alias{create_athena_query}
\alias{delete_athena_query}
\title{AWS Athena Queries}
\usage{
list_athena_queries(n = 50, workgroup = NULL, token = NULL, ...)

get_athena_query(id, ...)

create_athena_query(
  query,
  name,
  database,
  workgroup = NULL,
  description = NULL,
  ...
)

delete_athena_query(id, ...)
}
\arguments{
\item{n}{An integer specifying the maximum number of results to return.}

\item{workgroup}{A character string containing the workgroup's name.}

\item{token}{A character string specifying a continuation token, for pagi
nation.}

\item{\dots}{Additional arguments passed to \code{\link{athenaHTTP}}}

\item{id}{A character string containing an Amazon Athena Query ID. A vector of IDs can also be specified to request multiple queries.}

\item{query}{SQL that comprises the query.}

\item{name}{Name of the query.}

\item{database}{Name of the database to associate with the query.}

\item{description}{Description of the query.}
}
\description{
Get, Create, and Delete Athena Queries
}
\references{
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_ListNamedQueries.html}{API Reference: ListNamedQueries}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_CreateNamedQuery.html}{API Reference: CreateNamedQuery}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_GetNamedQuery.html}{API Reference: GetNamedQuery}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_BatchGetNamedQuery.html}{API Reference: BatchGetNamedQuery}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_DeleteNamedQuery.html}{API Reference: DeleteNamedQuery}
}
