% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/athena_workgroups.R
\name{list_athena_workgroups}
\alias{list_athena_workgroups}
\alias{get_athena_workgroup}
\title{AWS Athena Workgroups}
\usage{
list_athena_workgroups(n = 50, token = NULL, ...)

get_athena_workgroup(workgroup, ...)
}
\arguments{
\item{n}{An integer specifying the maximum number of results to return.}

\item{token}{A character string specifying a continuation token, for pagi
nation.}

\item{\dots}{Additional arguments passed to \code{\link{athenaHTTP}}}

\item{workgroup}{A character string containing the workgroup's name.}
}
\description{
Get Athena Workgroups
}
\references{
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_ListWorkGroups.html}{API Reference: ListWorkGroups}
\href{https://docs.aws.amazon.com/athena/latest/APIReference/API_GetWorkGroup.html}{API Reference: GetWorkGroup}
}
